app/requirements.txt
alabaster==0.7.12
alembic==1.8.1
annotated-types==0.7.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asn1crypto==1.5.1
asttokens==2.0.8
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
backports.shutil_which==3.5.2
bcrypt==4.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.7.0
cachetools==5.2.0
cattrs==23.1.2
certifi==2024.8.30
cffi==1.17.1
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
colorclass==2.2.2
comm==0.2.2
commonmark==0.9.1
configparser==5.3.0
coverage==6.4.4
cryptography==43.0.1
cssmin==0.2.0
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
dnspython==2.6.1
docopt==0.6.2
docutils==0.17.1
dominate==2.7.0
email-validator==2.1.0.post1
entrypoints==0.4
enum34==1.1.10
esbonio==0.14.0
exceptiongroup==1.2.0
executing==1.0.0
fabric==3.2.2
factory-boy==3.2.1
Faker==14.2.1
fastjsonschema==2.16.2
Flask==3.0.3
Flask-Assets==2.1.0
Flask-BabelEx==0.9.4
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
flask-nav==0.6
Flask-Principal==0.4.0
Flask-Security-Too==5.5.2
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
fmt==0.3.1
FormEncode==2.0.1
funcsigs==1.0.2
google-api-core==2.19.2
google-api-python-client==2.62.0
google-auth==2.34.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
googleapis-common-protos==1.56.4
greenlet==3.1.1
gunicorn==23.0.0
html2docx==1.5.0
httplib2==0.20.4
idna==3.10
imagesize==1.4.1
importlib-metadata==4.12.0
importlib_resources==6.4.4
invoke==2.2.0
ipaddress==1.0.23
ipython==8.21.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.5
jsmin==3.0.1
jsonschema==4.16.0
lazy==1.5
loutilities==3.9.6
lsprotocol==2023.0.1
lxml==5.3.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
mock==4.0.3
more-itertools==8.14.0
mysqlclient==2.2.4
nest-asyncio==1.5.5
nose2==0.12.0
oauth2client==4.1.3
oauthlib==3.2.2
packaging==21.3
pandocfilters==1.5.0
paramiko==3.5.0
parso==0.8.3
passlib==1.7.4
pathlib2==2.3.7.post1
pbr==5.10.0
pickleshare==0.7.5
pip-upgrader==1.4.15
pipdeptree==2.3.3
platformdirs==4.1.0
prometheus-client==0.14.1
prompt-toolkit==3.0.43
proto-plus==1.25.0
protobuf==4.25.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
pygls==1.3.1
Pygments==2.18.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
pyspellchecker==0.7.0
python-dateutil==2.8.2
python-docx==0.8.11
python-editor==1.0.4
python-slugify==6.1.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==26.2.0
QtPy==2.2.0
requests==2.32.3
requests-oauthlib==1.3.1
rich==13.9.4
rsa==4.9
rstcheck==6.2.4
rstcheck-core==1.2.1
scandir==1.10.0
Send2Trash==1.8.0
setuptools==75.6.0
shellingham==1.5.0
simplegeneric==0.8.1
singledispatch==3.7.0
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
speaklater==1.3
Sphinx==5.1.1
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sphinxcontrib-websupport==1.2.4
SQLAlchemy==1.4.51
sqlalchemy-datatables==2.0.1
SQLAlchemy-Utils==0.38.3
stack-data==0.5.0
terminado==0.15.0
terminaltables==3.1.10
testpath==0.6.0
text-unidecode==1.3
tinycss2==1.1.1
tornado==6.4.1
traitlets==5.4.0
typeguard==4.4.1
typer==0.12.5
types-docutils==0.18.3
typing==3.7.4.1
typing_extensions==4.12.2
tzdata==2022.2
tzlocal==4.2
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==2.2.3
validators==0.20.0
visitor==0.1.3
wcwidth==0.2.5
webassets==2.0
webencodings==0.5.1
Werkzeug==3.1.2
widgetsnbextension==4.0.3
win_unicode_console==0.5
wrapt==1.17.0
WTForms==3.0.1
zipp==3.20.2
